+++ relay-server/src/endpoints/attachments.rs
@@ -1,49 +1,43 @@
-use axum::extract::{DefaultBodyLimit, Multipart, Path};
+use axum::extract::Path;
 use axum::http::StatusCode;
 use axum::response::IntoResponse;
-use axum::routing::{post, MethodRouter};
-use relay_config::Config;
+use multer::Multipart;
 use relay_event_schema::protocol::EventId;
 use serde::Deserialize;
 
 use crate::endpoints::common::{self, BadStoreRequest};
 use crate::envelope::{AttachmentType, Envelope};
-use crate::extractors::RequestMeta;
+use crate::extractors::{Remote, RequestMeta};
 use crate::service::ServiceState;
 use crate::utils;
 
 #[derive(Debug, Deserialize)]
-struct AttachmentPath {
+pub struct AttachmentPath {
     event_id: EventId,
 }
 
 async fn extract_envelope(
-    config: &Config,
     meta: RequestMeta,
     path: AttachmentPath,
-    multipart: Multipart,
+    multipart: Multipart<'static>,
 ) -> Result<Box<Envelope>, BadStoreRequest> {
-    let max_size = config.max_attachment_size();
-    let items = utils::multipart_items(multipart, max_size, |_| AttachmentType::default()).await?;
+    let items = utils::multipart_items(multipart, |_| AttachmentType::default()).await?;
 
     let mut envelope = Envelope::from_request(Some(path.event_id), meta);
     for item in items {
         envelope.add_item(item);
     }
+
     Ok(envelope)
 }
 
-async fn handle(
+pub async fn handle(
     state: ServiceState,
     meta: RequestMeta,
     Path(path): Path<AttachmentPath>,
-    multipart: Multipart,
+    Remote(multipart): Remote<Multipart<'static>>,
 ) -> Result<impl IntoResponse, BadStoreRequest> {
-    let envelope = extract_envelope(state.config(), meta, path, multipart).await?;
+    let envelope = extract_envelope(meta, path, multipart).await?;
     common::handle_envelope(&state, envelope).await?;
     Ok(StatusCode::CREATED)
-}
-
-pub fn route(config: &Config) -> MethodRouter<ServiceState> {
-    post(handle).route_layer(DefaultBodyLimit::max(config.max_attachments_size()))
 }